eliminate fault in graphicAt()
making window management more robust
replacing more doubles with floats

+++ primary/GraphicsDecoder.hpp
@@ -107,7 +107,7 @@ public:
             }
             break;
             case GCMD_STROKEWEIGHT: {
-                fGraphics->strokeWeight((double)readFloat());
+                fGraphics->strokeWeight(readFloat());
             }
             break;
